Replace AdapterManager call of handleNewThing with listener
The Actions model uses a one-shot promise to wait for AdapterManager to add
new things. What this means is that if an Adapter adds more than one new
Device in response to startPairing, the user will only be able to see
the first Device in the Add New Things menu. This is because the Actions
model only calls Things.handleNewThing on the first Device the
AdapterManager finds (which resolves the Promise returned by
addNewThing).

Instead of this one-shot promise approach, I attach a listener for the
'thing-added' event to the AdapterManager which properly sends all new
things over the websocket connection to the UI.