Increase integration testing; Fix spec compliance of ActionsController (#108)
* Send raw newThing in handleNewThing instead of a sanitized Thing

* Replace AdapterManager call of handleNewThing with listener

The Actions model uses a one-shot promise to wait for AdapterManager to add
new things. What this means is that if an Adapter adds more than one new
Device in response to startPairing, the user will only be able to see
the first Device in the Add New Things menu. This is because the Actions
model only calls Things.handleNewThing on the first Device the
AdapterManager finds (which resolves the Promise returned by
addNewThing).

Instead of this one-shot promise approach, I attach a listener for the
'thing-added' event to the AdapterManager which properly sends all new
things over the websocket connection to the UI.

* Add test of /new_things and ws://.../new_things

* Use server over app in testing

* Add Istanbul-based Coverage

Coverage reports can be generated with `npm run cov` or
`npm run cov-html`.

* Remove ws.on('message') listener

* Add tests for thing pairing and thing removal

* Convert getActions to return an Array

The WoT specification states that `/actions` should return a list of
action objects.

Fix #105

* Delete actions with invalid names

Fix #104

* Add tests of ActionsController

* Test device removal

Fix #93

* Add unpair test

* Add nonexistent device unpair test

* Remove log statement

* Add test of removing a nonexistent action

* Delegate to getter functions in asDict

* Add AdaptersController tests

* Remove dead /:adapterId/:action route

* Add AdaptersController 404 Test

* Add tests of retrieving specific actions

* Update use of assert

* Update adapters-test to new style

* Remove all existing actions at the start of test

* Trim blank lines

* Replace Promise wrapper with chai-http's Promise support

* Switch to async/await style

* Ensure failing requests will fail tests