Use fetch in commands_controller; Remove request-promise dependency
For some reason, having request-promise installed poisons bluebird's
promisified request function in such a way that if the requested URL
returns a 400 status code a StatusCodeError appears in a completely
isolated stack. This code somehow bubbles up to Jest, causing it to fail
the tests.

This is the weirdest JavaScript bug I have ever seen.