Fix issues with caching static files.
The gateway was setting a max age of 14d for static resources.
However, it also sets a weak ETag header by default, which is
better for caching purposes when we're changing files, so we'll
now just rely on the ETag.

I tested this and verified that resources were properly cached
until modified.

Fixes #454