Open internal logs in same window
On touch screen kiosk systems there is no tab management. If a user opens the internal logs, they get 'stuck' in a new window. 
Swiping from the left to go back doesn't work in a fresh tab. But it would work if the logs were opened in the same window.