Open license in same window
For touch screen kiosk users, opening a new tab can lead to complications. The 'back' button no longer works. And since the license is shown as a text file, there is no 'close' button to close the new tab.

If the license if shown in the same tab, then the 'swipe to go back' funtionality still works, and users can get back to the UI.

Perhaps an even more optimal solution would be to load the license in an overlay.